Add quit game option
In order to add the game quit, a option to finish the turn had to
be created. This makes it easier to understand how to deal with
the menu, in order to continue playing the game.

When the game has been quit, the player dies, which makes the game
follows its ordinary ending flow.

Closes: https://github.com/henriquesimoes/heroquest/issues/63