Add argument to change game interface
Now, the graphic interface is the default launching choice.
If the command line interface is desired, the "--cli" option has
to be given to the game start-up class.

In order to make the command line parsing, the org.apache.commons
for command line interface package is being used. In order to make
the dependencies available to the code respository, all IntelliJ
configuration files have been included.